Compute rvalue bounds for more expressions. (#167)
This adds computation of rvalue bounds for the following pointer-typed expressions:
- Pointer addition and subtraction.
- Post/pre-increment and decrement.
- Assignments (assignments produce rvalues in C).
- Compound assiginments involving addition and subtraction.
This completes Github issue #121.

In addition, it adds computation of rvalue bounds for the following integer-typed expressions (section 5.4.2 in the the Checked C version 0.6 specification):
- Addition, subtracation, and multiplication.
- Post/pre-increment and decrement.
- Bitwise logical operations
- Unary +, -, and bitwise complement.

During testing, I found that the code for checking that compound assignment operators (+= and so on) have bounds was not correct and triggered an assert..  The code was just checking the bounds of the rhs expression.  We actually  need to check the rvalue bounds of the entire expression for compound assignment operators. For pointer arithmetic, this is a no-op (there's no possible way it could fail to have bounds), but for integer arithmetic, it is not a no-op.

Testing:
- Added new tests for inference of pointer arithmetic bounds.
- Still need to write tests for bounds-inference for integer-typed expressions.
- Passed Checked C regression tests.
- Passed clang regression tests.