Remove Alignment Specifiers to make tests cross-platform (#172)
These tests previously included verification of the "align n" specifiers on loads and stores. This is type-size/alignment dependent, which of course is therefore platform-dependent, causing tests to fail on Linux and Mac. 

I have removed them, because the test is really about the order of loads/tests/stores around bounds checks, and type-alignment is not something we are yet focusing on.