Adds Unix (Linux) & Mac Build Automation (#178)
This allows us to build Checked C (including the large test suite) on Unix (Linux) and Mac. Building on Mac is very similar to building on Unix, with a few overrides. 

Scripts can be run in any directory. They will treat that directory as the location to check out the source and to create a build dir within.

At the moment this only builds master branches for all repos. Later we can add the functionality to customise this. 

The scripts download and install CMake if the system CMake is not version 3.7, or if the previous download is not still around. This is done automatically. Build directories are not deleted after building, to speed up subsequent builds.