Fix file race condition involving driver tests.
Nightly testing failed with an error that looks like a race condition in file
access. There were several RUN lines that all output the same file. One of
them failed with an error that it was unable to open the output file.  This
is tracked by issue #181.

The root case is that we have several test files that only differ in their
file extension, which is based on the programming language.  We have
checkedc-notsupported.c and checkedc-notsupported.cpp, for example.  We
do not specify an output file, so the default checkedc-notsupported.o is
used.   The tests are executed in parallel and can collide.

This changes the RUN lines to use a temporary files for the output file