Eliminate separate build and test steps in automation. (#187)
We are seeing failures in automated builds on Windows due to failing file copies.   We are doing back-to-back invocations of the build system - one to build clang and one to build and run tests.   The second invocation is failing with a problem writing files that are also written by the first invocation.  The reason for the failure is unclear, but a simple workaround is to only build and run tests.  That automatically invokes building clang and the duplicate writes go away.

Details of the problem:
At the end of building clang, the clang build always create a few binaries with the same name as existing compilers.  It does this even if nothing changes during the build.  On UNIX, the build always creates new symbolic links.  On non-UNIX platform, the build copies the clang binary to the new file names.    The first build invocation builds clang, which copies those files.  The second invocation to build and run tests always does a clang build as its first step, which also copies those files.  Those copies are what is sporadically failing.   It is not obvious why the copies are failing.  The copies are being made using `cmake -E copy`, which makes this hard to debug.    It is nice to have separate build and test steps in automation, but we do not need them.  Since this seems to be causing problems, we'll stop doing this.