Add scripts for automated build and test on Linux. (#192)
This change adds scripts for automated building and testing of the Checked C version of clang on Linux using VSTS.   The preliminary scripts for UNIX are moved to a travis subdirectory for possible future revision.

The scripts are organized the same way that they are for Windows.    We have one script that does everything.  It invokes separate scripts to check and configure environment variables, to set up files, to invoke cmake, and to build and test clang.

Testing:
-  VSTS automated job passes on a Linux x64 box running Ubuntu 16.04.
-  The X86 target is not tested yet.