Fix automation for checking out branches on Windows. (#194)
The commands for checking out branches were not right and could cause
merges when branches were specified explicitly.  Use the same commands
that we use for the Linux scripts.  We don't have functions, so just
place the commands inline.

Also make the clone commands more resilient in the Windows scripts.  We only
checked for the presence of the LLVM source directory, so if we something went
wrong with the clones of clang or Checked C, we would not clone them the next
time they ran.  Follow the sasme approach used in the Linux scripts where
each the clone of each repo is checked separately.

Print out the BINARIES_DIRECTORY variable at the end of the config-vars
script, in case it was set incorrectly.

Testing:
- Ran automatation scripts locally on Windows.
- Built and run the modified config-vars script for Linux under the
  Windows Subsystem for Linux.