Do not generate error message 'expected bounds expression' when there is a semantic error (#197)
Tests for rel_align show that sometimes we generate the error message "expected bounds expression before relative bounds clause" when there is a typechecking error in the bounds expression. We should not generate this message because it will confuse programmers. We should avoid generating this error message when the bounds expression is invalid.  This address issue #195