Rewriter uses bounds information to add constraints for _Ptr types (#227)
The rewriter can now check for bounds safe interfaces and update the constraints related to _Ptr types based on what it finds. Right now, it adds constraints that are not understood by the constraint solving algorithm, and making the constraint solving algorithm understand them is next. It only supports _Ptr types, but can support arrays later in the same style.