Fix compiler assert during bounds inference. (#232)
The LLVM nightly test MultiSource/Benchmarks/McCat/09-vor is failing with an
internal compiler error during bounds inference.

The problem is that we were assuming that all base expressions of structures
are lvalues. According to the C specification, they do not have to be lvalues.
For example, a call can produce a structure value.  The result of a call is
not an lvalue. There can be a member reference on the value, for example:
  g().f;

The fix is to check that base expressions are lvalues before trying to
add bounds checks.

Testing:
- Added a new regression test that shows the issue.
- Passed Checked C regression tests.