Fix clean builds in Windows for automated testing.
Visual Studio Team Services clones the checkedc-clang repo and then runs
scripts from that repo.  To handle the multi-repo set up for Checked C,
the scripts create clones of the Checked C repos in the sources directory.
The VSTS automation does not know about these clones, so it doesn't remove them
when a clean build is requested.  It is using a git command to clean the
repo from which automation is run.

This change modifies the automation scripts to remove the additional repos
when a clean build is requested.  The scripts also remove the output directory,
if it exists, when a clean build is requeated.

The variable BUILD_CHECKEDC_CLEAN control whether cleaning is done.  It
can be set to "Yes" or "No".  The user can directory set the variable.
If it is not set, the script checks whether the BUILD_CLEAN variable
is defined to set the variable.  The VSTS system defines BUILD_CLEAN
when a clean build is needed.

This problem was found when an automated build on Windows became corrupted
because the linker ran out of memory during an incremental link.  Requesting
a clean build did not produce the expected results.