Elevated execution bug (#255)
It turns out windows hates 32-bit executables which contain the words "install", "setup" or "update" in their filename, and requires they run with elevated permissions. https://msdn.microsoft.com/en-us/enus/library/aa905330.aspx http://stackoverflow.com/questions/11573444/why-is-windows-asking-for-system-administrator-privileges-for-running-executable

This means when we compile and run some tests with "update" in the name, through Visual Studio, the tests fail as it can't execute the files concerned. This PR just renames the files and adds some documentation.