Pointer_bounds_cast.c test fails on Windows 86.
The new pointer_bounds_cast.c test failed with an internal compiler assertion
during automated compiler testing. The problem is  the children method in the
BoundsCastExpr class. The methods returns an iterator. The iterator was being
built by taking the address of local variable. When the method returned, the
iterator structure pointed to a stack location that had been deallocated.

The fix is to have the iterator structure point to heap-allocated memory
that is part of the AST.  This required refactoring the way bounds expressions
are stored in cast expressions. Both it and the op expression need to be
stored in an array of Stmt *, following the pattern used elsewhere in the
AST for children of AST nodes.  I moved the BoundsExpr class earlier in
the file to avoid a forward reference to a BoundsExpr, which would have
required a reinterpret_cast.

Testing:
- Passed Checked C regression tests.
- Clang regression tests are running now.