Infer bounds for struct member accesses. (#258)
Infer bounds for uses of struct members that have member bounds declarations or that are checked arrays.  The main implementation challenge is that the member bounds have to be transformed from using members to using member accesses.  We use a tree transform class to do this.  The address issue #249.

This change implements bounds inference for rvalues produced by member references and bounds inference for lvalues produced by member inferences.

-  Struct member accesses in contexts that expect rvalues will produce rvalues. These rvalues need bounds.   An example of this is:
```
struct S {
  array_ptr<int> p : count(len);
  int len;
}

struct S tmp;
array_ptr<int> p : count(tmp.len) = tmp.p;  // need bounds for tmp.p
```
- In other cases, we need the bounds for the lvalues produced by struct member accesses.   Examples of this are:
```
struct  S {
   int a checked[5];
}

struct S tmp;
tmp.a[3] = ...  (we need the bounds for tmp.a).
```
or
```
struct S { 
  int f;
}

struct S tmp;
  ...   &tmp.f  ...
```
In the first case, the reference to `a` produces an lvalue that is converted from array type to pointer type, but not dereferenced.

Testing:
- Add clang test of inference of bounds for reads of structure members and assignments to struct members.
- Update existing clang tests that use structure members to expand the information from 'Any' to
  the actual bounds.
- Add dynamic test to the Checked C repo that reads through a pointer stored in a structure member.
- More comprehensive testing is needed.    This changes unblocks other people, so committing it now.