Halve the parallelism for Windows automated test builds. (#267)
The automated builds are timing out on Windows. VSTS reports that the agent "did not renew the lock", i.e. the agent is not calling home enough. One possibility is that we're having network issues. Another possibility is that the switch to the x64 toolset increased memory usage, caused the agent machine to page. For now, just halve the parallelism while I try to sort this out.