Fix usage of member bounds with the arrow operator.
This addresses issue #280, found when converting some benchmarks in LNT
to use Checked C.  Two benchmarks that used the arrow operator to
access members with bounds failed to compile because they caused
internal compiler asserts.   The problem was that the conversion of
member bounds to concrete instances of bounds on variables and expressions
was incorrect.  The concretization function neglected to take into account the
arrow operator, even though the the kind of member access operator had been
passed in to it.

Testing:
Added several small tests to make sure that the feature works in some
cases so that others blocked on this bug can make forward progress.
Additional tests need to be written.

New tests in the checkedc-clang repo are:
- Add a test of dumping bounds for structures in the AST
(ast-dump-struct-bounds.c), so that we can be sure that the input
to the concretization function is reasonable.
- Add tests of the inferred bounds for members with member bounds,
where the members are accessed using the arrow operator
(member-arrow-reference.c).

Also added runtimes test to the Checked C repo in
tests\dynamic_checking\read-member-arrow-ponter-check.c.
This is the read-member-pointer-check.c adapted to
use pointers to structures.