Correct CL_MPCount variable in documention. (#293)
We accidentally used the name CL_CPUCount for a build environment variable that controls the number of parallel C compiles that will be forked by the C compiler driver. The correct name is CL_MPCount. CL_CPUCount has no effect, which leads machines to suffer from VM thrashing when msbuild is invoked by hand following directions in the documentation.