Check types of bounds_cast operands and allow bounds expressions in bounds_cast operators.
For bounds_cast operators, check that the first expression has type that is suitable for conversion to a pointer type.   It should have a pointer type or an integral type.

Let bounds_cast operators take a bounds expression that specifies bounds as their second argument,
For example, we now allow `count(10)` or `bounds(a,b)` as the second argument for a bounds_cast operator.  This makes code easier to read and understand.

Testing:
- New tests have been added to the Checked C repo.