Fix typo in Windows automation script.
For each Checked C repo used in automation, we have a variable that controls
which branch is used during a test run.   This allows us to test changes
that span repos.  We can point the automation at specific combinations of
branches.  We were using the wrong branch name for the checkedc-clang repo in
the Windows test scripts.  We are using CHECKEDC_BRANCH when we should have
been using CLANG_BRANCH. I checked and this bug did not exist on for the
Linux bash scripts.