Type-checking error leads to compiler crash (#296)
A type-checking error removes the lvalue attribute on a MemberExpr. This causes an assert in SemaBounds to fail and the compiler to crash.   This changes the assert to be a conditional. Although there are numerous lvalue assert checks, it looks like that assert is the only assert that comes across this problem.  This fixes #291.

Testing:
- Added a new test case that makes sure the compiler does not crash in in the presence of this type checking error.