Implementation-specific definitions of NULL in checked scopes.
The C standard allows NULL to be an implementation-defined null pointer
constant.  The clang implementation of C allows NULL to be defined as
((void *) 0).   This is rejected by the static checking of checked scopes
currently because it is a cast to an unchecked pointer.   However, we would
like to allow NULL to be used in checked scopes.

The solution is to allow null pointer constants to appear in checked scopes,
however an implementation defines them.  This changes modifies the checking
of casts in checked scopes to allow ((void *) 0).  We may need to do some
extra work to preserve the safety property of checked scopes. You can't
do pointer arithmetic because void is an incomplete type or convert it some
other unchecked pointer type in a checked scope.  You may be able to try
to dereference it.  This could lead to undefined behavior, which could
lead to safety problems given how C compilers optimize in the presence
of undefined behavior.

Testing:
- Add a new clang-specific test for its implementation defined null pointer
  constants.
- Passes existing Checked C tests.