Revisit function to pointer type decay in a checked scope (#304) (#308)
Revisit function to pointer type decay in a checked scope (#304)

- Do not allow functions with old-style K&R parameter lists to be used in checked scopes.
- After fixing that, in checked scopes, change the implicit conversion from function type to pointer-to-function type  to be produce checked pointers again.

In checked scopes, use or declaration of no-prototype functions is nota llowed.  However, functions declared with old-style K&R parameters list are a little bit exceptional.   K&R parameter lists have a special form where function parameters re declared outside of function definition, so there is type information at the declaration.  However, when it is used, it is treated as a no-prototype function. 

A  DeclRef expression is built for a function call.  If the call target is a K&R parameter function,  the code generates a no-prototype function DeclRefExp.  In short, even though functions with K&R-style parameter lists are  declared in ways similar to prototype function declarations,  at call sites, they are treated as no-prototype functions.   In a checked scope, this no-prototype function call should be prevented.

C implicit converts uses of functions from function type to pointer to function type.  In checked scopes, this conversion used to produce a checked function pointer type.   In a recent change, it was changed to be produce an unchecked function pointer type.   After fixing the problem with K&R function declarations. restore this conversion.