Copy Checked C headers to clang system header directory. (#317)
The Checked C header files are stored in the Checked C repo.  If the Checked C repo is in the tree, add a dependency on the target that copies the Checked C headers.   The target is conditionally added by the LLVM cmake files and a variable is set for use by the clang build system.