Fill in lexicographic comparison functions.
Fill in the lexicographic comparison functions for expressions and
declarations.  This passes all the existing tests in the Checked C repo
and all but one of the Checked C tests in the clang repo.

Lexicographic comparison functions for types still need to be
implemented.  There are also some expressions that still need to be
filled in.