Create lexicographic ordering of C expressions (part 1) (#335)
This implements a lexicographic ordering of C expressions and switches comparison of bounds expressions to use it.  The lexicographic ordering provides a total order on C expressions.

We order by expression kind first  We then order by properties of expressions after that.  For expressions that have an identical kind and properties, we use the lexicographic order of their subexpressions.
-  The ordering is modulo parameter names and typedefs.  For parameters, we use the positional index of the parameter and its type in the ordering, and ignore the names.  For typedef type names, we use the canonical type name, where typedef'ed have been replaced with their underlying type.
- For non-parameter variables, we use their name, type, kind of scope, and order of declaration within a scope for the order.
- For constants, we use the value and type of the constant.

Some aspects of the lexicographic order have yet to be implemented.  The main one that is missing is a lexicographic order on types.   We determine equality/inequality properly for types by deferring to the clang type system APIs for that, but do not order types properly.   There are a few operators that have not been implemented - these are marked with TODOs.

This means that the lexicographic ordering can only be used reliably for equality/inequality of expressions at this point.

Testing:
- Adding a new test of lexical expression equality/inequality indirectly using redeclarations of bounds expressions to the Checked C repo.
- Passed existing Checked C and clang tests.
- Passed LNT test automation for Checked C, which includes benchmarks modified to use Checked C.