Checkpoint work on checking bounds declarations modulo variable equality.
- Prepare for analyzing which variables are equal.
- During check of bounds declarations, build the CFG, but don't use it.  The
  Checked C tests pass.
- Start writing analysis for conservatively determining which variables are
  equal at program points.
-- Design interface for analysis.
-- The general approach will be to do an intraprocedural analysis of a
   functio nbody.  We will  then iterate through the basic blocks and walk
  through the statements in each basic block to determine the equivalance
  sets for variables.
-- Write some support code for the analysis (this codie compiles, but is
  not complete and is untested).

The support code includes a class for (space)-efficiently representing small
equivalence sets of integers. We stiill need to optimze the code for look up.
We can't use the existing clang class for integer equivalence sets because
that provides a union/find API.  We want the dual of union/find, which
is partition refinement.

The support code also includes classes for determining which variables are
address-taken and expressions that have invariant values during the execution
of function bodies.  The code for determining invariant values is incomplete.

Testing:
- The code compiles and Checked C tests pass.