Type variable type. (#340)
This pull request has TypeVariableType implementation.
When a generic function is declared, each type variable is assigned a TypeVariableType, and is stored in Typedef. At the same time, each Typedef is stored inside of FunctionDecl to keep track of the name (identifier) and the type (deBruijn indexes).
Function Type is also updated to keep track of the number of type variables.
When invoking -ast-dump, the type variable information is shown as well.
The Typedef that stores TypeVariableType is shown first.
Then, the Function declaration is shown with the new function type. The new function type shows the number of type variables declared in the function.
Next, each TypeVariable will be shown.
Then the list continues like it was before, starting with the list of parameters.