Allow K&R style function definitions preceded by prototypes in a checked scope. (#357)
We disallow function declarators that have no prototypes in checked scopes because calls to them may not be typechecked.  We were being too strict in the checking and  just looking at whether a function declarator provided a prototype. 

We need to take into account that according to the C11 specification, type information from prior declarations of a function is merged with a function declarator to determine whether a function has a  prototype.    This change moves checking to later in ActOnFunctionDeclarator after type merging, which addresses issue #355.  If a no-prototype function declarator is preceded by a prototype, the declarator is considered to have a prototype.  

However, we need to add some special-case handling for the case of K&R style function definitions. 
This is where the function declarator for a function definition is followed by a list of declarations for parameters, for example:

int f(a, b)
int a;
int b;
{ }

clang assigns a prototype function type to this definition, even though the C11 standard does not say that such a definition creates a prototype for a function.  We could have issues with other C compilers in the future, which may not type check succeeding uses of the function. We specifically check that there is a prior declaration with a prototype.  There may be an easier way in clang to determine that a K&R function definition has a prototype in a standard-compliant way (because of a prior declaration),  but I couldn't find it.

Even though K&R-style function definition are obsolete, we care about this because we are trying to minimize code changes required to use Checked C.  We don't want to force someone to modify a K&R-style function definition when they have done their homework and provided a prototype.

This change also includes some minor clean up and improvements:
- We now record whether a function declarator has a checked/unchecked specifier.
- There were some identifiers that started with lowercase letters  that needed to start with uppercase letters (and vice versa) to match the coding convention.
- Make some comments more accurate.

Testing:
- I added a new test case to the Checked C tests for a K&R function definition preceded by a prototype.  This was covered by a separate pull request for the Checked C repo.