Extend -Wcheck-bounds-decls to variable declarations with initializers (#358)
Checking the correctness of bounds declarations is currently under a warning flag that is off by default (-Wcheck-bounds-decls).    This is because the checking that declared bounds are implied by inferred bounds is not very powerful.  It is limited to just syntactically identical bounds declarations and special cases for bounds(none) and bounds(any).    There'd be lots of spurious warnings if it was on by default.

The checking of bounds declarations is done only for assignments.  This extends it to variable declarations with initializers.   Most of the work went into producing good warning messages that have notes explaining what is going on.   Assignments were using expanded bounds declarations (where count/byte_count have been expanded into their underlying range expressions).   I implemented new logic to create expanded bounds declarations for variable declarations with bounds.   This is so that the contents of notes match up.  I had to make some methods public on BoundsInference.  It isn't clear that the methods are in the right place, but I've left that code refactoring for another day.

Other than that, I refactored the logic for checking bounds declarations for assignments so that we could re-use it for declarations with initializers.

Testing:
- Passed existing Checked C tests.
- Passed clang test run..
- Added a new clang-specific tests that checks bounds inference for initializers.  Also check the notes generated with warnings.