Remove Ptrtoint from Dynamic Check CG (#360)
It turns out you can icmp pointer types in LLVM IR, without converting them to integers first. This pull request removes our generation of "ptrtoint" instructions.

However there is an issue in inference, ICMP needs both operandi to have the same type, and for complex reasons sometimes the type of the bounds doesn't match the types of the pointer we're comparing to. This means sometimes we have to insert a bitcast in the LLVM IR. Semantically this keeps the same value but changes the IR type to match what is needed. I need to add code comments to clarify how we use these casts.

I also cleared up the method name used for generating dynamic_bounds_cast so it doesn't overload that of the dynamic bounds check, and split out statistics for dynamic checks from casts to be desperate to bounds checks.