Use bounds-safe interfaces on function pointers during type checking (#363)
Most of the work for this change was revamping the error messages for uses of unchecked types in checked scopes.    It is unclear what the term “unchecked type” means when constructed types are involved.  Constructed types are types such as function types, array types, and pointer types.  The problem may be a type that a type is constructed from (for example, a function type involving a parameter with a unchecked pointer type) or it may be in a bounds-safe interface.  In addition, the problematic unchecked type may be unclear. 

We revamped the error messages to be clearer and added notes explain what is going on.  We:
- use a simple error message for simple errors such as trying to declare or use a variable or parameter with an unchecked pointer or array type in a checked scope
- use a longer error message when the problem is in a type that a type has been constructed from.
- use a different error message when the problem is the use of bounds-safe interface with an unchecked type in a checked scope.

We customize the error message based on whether the problem is a definition or a use.  For a use, we add a diagnostic note that points out the definition.   If it is not obvious what the problematic unchecked type is, we add a diagnostic note with the problematic type. We do this when the problem is a type that the type was constructed from, a type occurring in a bounds-safe interface, or the type is a type definition.

Testing:
- There will be a separate pull request for the Checked C repo the updates the checked scope tests for the new error messages.
- Add clang-specific tests of the notes and error messages for uses of unchecked types in checked scopes.
- Add tests of AST dumps with bounds-safe interface types, since those were missing.
- Passed existing Checked C and clang tests.