Conversions involving checked void pointers (#365)
This addresses GitHub issue #364.   It allows any kind of checked pointer to be implicitly converted to any kind of checked void pointer.   Previously we required that the target and destination types be the same kind of checked pointer.  

Also disallow conversions between void pointers and checked function types.   In examining the code for conversions involving void pointers, I noticed that the code allowed conversions between void * and function pointers as an extension.   This would also allow conversions between checked void pointers and function pointers.  We do not want to allow this, even if the static analysis for ensuring that checked function pointers are valid would also disallow this.

Improve the comment about conversions between different checked pointers to the same referent type.

Testing:
- Update tests in the Checked C repo to allow the conversions to different kinds of checked void pointers.  This is covered by a separate pull request.
- Add  tests that conversions between void pointers and checked function pointers are not allowed.
- Passes existing Checked C and clang tests.