Add missing cast for bounds-safe interfaces (#367)
At bounds-safe interfaces, we allow values with checked pointer types to be assigned to lvalues with unchecked pointer types.  We also allow values with checked pointer types to be passed to arguments with unchecked pointer types.

We were missing a cast from the checked type to the unchecked type in the AST.  This triggered an assert during code generation for parameter passing because the type of the argument was incompatible with the type of the parameter.   We had a workaround that modified the assert to allow special behavior at a bounds-safe interface.

This change adds the missing cast and removes the workaround.   It modifies CheckSingleAssignmentConstraints to take the interop type as an additional argument.  If the match fails with the original LHS type, we try the initerop type.   There are a lot of special cases in CheckSingleAssignmentConstraints.  I was concerned that some of them would apply to the LHSInteropType.  However, in examining the special cases it seems that they all apply to non-C languages.   So the logic in CheckSingleAssignmentConstraints for handling the interop type can be simple.  We just need to check at the end of the function that the interop type works if the LHS type didn't work. 

If we use the interop type, we insert a downcast to the unchecked type. For now, I use a bitcast operator.   This should be changed to use a new cast operator for clarity in the AST.  I plan to make that change separately.

Testing:
- Added new test case from issue #143 that triggered the assert in the unmodified code.
- Passes automated testing on Windows and Linux.