Potentially Disallow Global Variable References in Parameter Bounds Expressions (#362)
This is preparation for a change we're thinking of making.

We cannot reason about how global variables change inside functions, which means our bounds static analysis may not be sound if global variables are allowed in parameter bounds. 

This adds some code to disallow global variables inside parameter bounds expressions. It guards it behind a flag that is currently false, because we're not 100% sold on needing this yet, and want to think about it more.

Testing:
- I set the variable `errorOnGlobals` to true, and ran the compiler on the YACR2 converted benchmark, and I got a lot of errors.