Cast lvalues in checked scopes with bounds-safe interfaces to checked types.
The Checked C specification calls for treated declared entities with unchecked
types and bounds-safe interfaces as though they have checked type when the
entities are used in a checked scope.

We are implementeing by this by inserting casts to checked types in the AST.
We do this as part of creating DeclRefExprs for uses of the declared values.
This commit adds casts for variable and member uses in checked scopes
that have unchecked types.

Still to do:
- Testing that the casts are inserted as expected and that bounds
  checks are inserted also.
- Rewriting function types with bounds-safe interfaces on parameters
and returns to use checked types.

Testing;
- Passes existing Checked C tests in the Checked repo.
- Passes existing clang-specific Checked C tests.